Git LFS instructions to CONTRIBUTING.md
If git lfs is not installed, the tests will not pass. This includes the unit tests. In addition, if LFS is installed after clone or the node version is not correct it can lead to test failures.

Updating the instructions to try to steer people to installing LFS and correctly validating node versions and LFS functionality prior to filing issues about tests failing.