tools: cache the built tools arrays unless it changes
We access '.tools' a lot and in https://github.com/microsoft/vscode/issues/252693 it seemed to cause issues for some users. Notably, we 'cloneAndChange' each tool definition each time it's called. This commit adds a cache to avoid recompuation as long as the array doesn't change. I think this is fine since a tool needs to re-register if it updates its description, schema, or other properties, which will trigger an update.