Enhance git commit prompt with repository context
The most important information is the branch name which can be helpful
to find out what the implemented feature is about or to extract ticket
numbers to be referenced in the commit message.

GitHub information about the repository owner/name and possible pull
request information is optional.

The implementation is based on [1].

[1] https://github.com/microsoft/vscode-copilot-chat/blob/aa45bf559489cc5d21e93bd5cce7388056819222/src/extension/prompts/node/agent/agentPrompt.tsx#L497-L524