tools: make mcp tool calls actually run in parallel
Closes https://github.com/microsoft/vscode/issues/271989

We didn't actually parallize calls in general because we need to include
prompt sizing information through our APIs, and so need to render each
'flexGrow' group sequentially. However, we know that MCP's don't consume
any of this information today, so we can do this calls in parallel safely.

Implementation isn't the prettiest but it does the job.